Add a Speech Service AudioProvider
Add an AudioProvider that sends data to the Microsoft Cognitive Speech
Services API. Because the Speech API expects silence to tell when a
sentence is finished we need to append some silence to the end of the
spoken text.

Unfortunately due to the way SRS is designed (Audio is buffered by
client but radio effects are by radio) this means we need to send a
terminator based on the radio in use rather than a specific client that
has finished speaking. This will mean that if two people are talking on
the same frequency we will have some issues with one being cut off but
this is an acceptable edge-case.