Switch UdpVoiceHandler UdpClient listener to async
Switch the UdpVoiceHandler udpClient to use `ReceiveAsync` instead of
`Receive`.

This allows us to start bringing Async into the application
which reduces manual thread management but the main benefit right now is
that by using async/await we no longer get Socket exceptions thrown when
disconnecting from a server due to `WSACancelBlockingCall`
interruptions on the `Receive` call.

This is one less exception to worry about and it means we can add
logging back in for when an exception is raised.

Another option that I didn't follow was to keep things as is and just
catch the SocketException and check the SocketErrorCode to see if it was
an interrupt or not (if not interrupt then log exception, otherwise
ignore).

The reason I didn't folow this path is that it still leads to
exceptions being thrown when we really don't need them to be.